+++ core/CHANGELOG.md
@@ -1,9 +1,14 @@
 # TBD
 * Small comment/doc changes from weekly review
 * Easy fix to sort `api_container_env_vars` alphabetically
+* Remove some now-unneeded TODOs
+* Fix the `build_and_run.sh` script to use the proper way to pass in Docker args
+* Disallow test names with our test name delimiter: `,`
 * Create an access controller with basic license auth
 * Connect access controller to auth0 device authorization flow
 * Implement machine-to-machine authorization flow for CI jobs
+* Bind-mount Kurtosis home directory into the initializer image
+* Drop default parallelism to `2` so we don't overwhelm slow machines (and users with fast machines can always bump it up)
 
 # 1.0.0
 * Add a tutorial explaining what Kurtosis does at the Docker level